Move motion control error code to dedicated module
